// backend/src/types.ts
export enum LayerParametersEnum {
  priority = 'priority',
  fill = 'fill',
  x = 'x',
  y = 'y',
  opacity = 'opacity',
  shape = 'shape',
  radius = 'radius',
  width = 'width',
  height = 'height',
  points = 'points',
  length = 'length',
  objects = 'objects',
}

type LayerParameters = {
  priority?: number
  fill?: string
  x?: number
  y?: number
  opacity?: number
  shape?: string
  radius?: number
  width?: number
  height?: number
  points?: Array<[number, number]>
  length?: number
  objects?: Identifier[]
}

export enum ObjectParametersEnum {
  priority = 'priority',
  fill = 'fill',
  x = 'x',
  y = 'y',
  opacity = 'opacity',
  shape = 'shape',
  radius = 'radius',
  width = 'width',
  height = 'height',
  points = 'points',
  length = 'length',
}
// ...
export enum AnimationParametersEnum {
  duration = 'duration',
  delay = 'delay',
  velocity = 'velocity',
  direction = 'direction',
  angularVelocity = 'angularVelocity',
  scaleFactor = 'scaleFactor',
  fill = 'fill',
  opacity = 'opacity',
}
// ...
export enum EntityTypes {
  object = 'object',
  layer = 'layer',
  animation = 'animation',
}

export type Execution = Step[]

export type Step = Persistence | Delete | Animate | Global | Conditional

// --

export type Persistence = Create | Update
// ...
export type Delete = {
  type: 'delete'
  identifier: string
  entity: EntityTypes
}

// --

export type Animate = {
  type: 'animate'
  animationIdentifier: Identifier
  entity: EntityTypes.layer | EntityTypes.object
  entityIdentifier: Identifier
  action: Action
}

export enum Action {
  start = 'start',
  stop = 'stop',
  speedUp = 'speedUp',
  slowDown = 'slowDown',
  reset = 'reset',
}

// --

type Global = Collision | Wait

export type Collision = {
  type: 'collision'
  value: boolean
}

export type Wait = {
  type: 'wait'
  duration: number
}

// --

export type Conditional = {
  type: 'if' | 'loop'
  condition: Condition
  count?: number
  body: Step[]
}

export type OP = '<' | '>' | '<=' | '>=' | '==' | '!='